ESP32C5/tests/requirements.txt
esptool==4.8.1
pyserial==3.5
pytest==8.3.4
